refactor(validator): extract SQL operations to persistence layer
* Create persistence/miners.rs with 9 miner table operations
* Create persistence/gpu_assignments.rs with 3 GPU tracking operations
* Create persistence/miner_nodes.rs with 2 node relationship operations
* Remove 13 SQL functions from verification.rs (~900 lines)
* Update all callers to use persistence layer methods
* Maintain transaction support for cleanup_gpu_assignments
* Add run_migrations() calls to integration tests for table setup
* Fix doc-test syntax errors (text/ignore blocks for examples)
