+++ crates/basilica-validator/src/miner_prover/verification.rs
@@ -5,7 +5,7 @@
 
 use super::miner_client::{MinerClient, MinerClientConfig};
 use super::types::MinerInfo;
-use super::types::{GpuInfo, NodeInfoDetailed, NodeVerificationResult, ValidationType};
+use super::types::{NodeInfoDetailed, NodeVerificationResult, ValidationType};
 use super::validation_states::{StateResult, ValidationState};
 use super::validation_strategy::{ValidationNode, ValidationStrategy, ValidationStrategySelector};
 use super::validation_worker::{ValidationWorkerQueue, WorkerQueueConfig};
@@ -574,41 +574,6 @@ impl VerificationEngine {
         Ok(nodes)
     }
 
-    /// Clean up GPU assignments for an node
-    async fn cleanup_gpu_assignments(
-        &self,
-        node_id: &str,
-        miner_id: &str,
-        tx: Option<&mut sqlx::Transaction<'_, sqlx::Sqlite>>,
-    ) -> Result<u64> {
-        let query = "DELETE FROM gpu_uuid_assignments WHERE node_id = ? AND miner_id = ?";
-
-        let rows_affected = if let Some(transaction) = tx {
-            sqlx::query(query)
-                .bind(node_id)
-                .bind(miner_id)
-                .execute(&mut **transaction)
-                .await?
-                .rows_affected()
-        } else {
-            sqlx::query(query)
-                .bind(node_id)
-                .bind(miner_id)
-                .execute(self.persistence.pool())
-                .await?
-                .rows_affected()
-        };
-
-        if rows_affected > 0 {
-            info!(
-                "Cleaned up {} GPU assignments for node {} (miner: {})",
-                rows_affected, node_id, miner_id
-            );
-        }
-
-        Ok(rows_affected)
-    }
-
     /// Helper function to clean up active SSH session for a node
     async fn cleanup_active_session(&self, node_id: &str) {
         // Direct SSH sessions are now managed at connection level
@@ -777,7 +742,8 @@ impl VerificationEngine {
             || node_result.validation_type == ValidationType::Lightweight
                 && node_result.ssh_connection_successful)
         {
-            self.cleanup_gpu_assignments(&verification_log.node_id, &miner_id, Some(&mut tx))
+            self.persistence
+                .cleanup_gpu_assignments(&verification_log.node_id, &miner_id, Some(&mut tx))
                 .await?;
             tx.commit().await?;
             return Ok(());
@@ -804,20 +770,22 @@ impl VerificationEngine {
                     miner_uid, node_result.node_id
                 );
 
-                self.ensure_miner_node_relationship(
-                    miner_uid,
-                    &node_result.node_id.to_string(),
-                    &node_result.node_ssh_endpoint,
-                    miner_info,
-                )
-                .await?;
+                self.persistence
+                    .ensure_miner_node_relationship(
+                        miner_uid,
+                        &node_result.node_id.to_string(),
+                        &node_result.node_ssh_endpoint,
+                        miner_info,
+                    )
+                    .await?;
 
-                self.store_gpu_uuid_assignments(
-                    miner_uid,
-                    &node_result.node_id.to_string(),
-                    &gpu_infos,
-                )
-                .await?;
+                self.persistence
+                    .store_gpu_uuid_assignments(
+                        miner_uid,
+                        &node_result.node_id.to_string(),
+                        &gpu_infos,
+                    )
+                    .await?;
 
                 // Create/update GPU profile for this miner after successful verification
                 let gpu_repo = GpuProfileRepository::new(self.persistence.pool().clone());
@@ -902,12 +870,13 @@ impl VerificationEngine {
                     miner_uid, node_result.node_id
                 );
 
-                self.update_gpu_assignment_timestamps(
-                    miner_uid,
-                    &node_result.node_id.to_string(),
-                    &gpu_infos,
-                )
-                .await?;
+                self.persistence
+                    .update_gpu_assignment_timestamps(
+                        miner_uid,
+                        &node_result.node_id.to_string(),
+                        &gpu_infos,
+                    )
+                    .await?;
             }
         }
 
@@ -923,1059 +892,9 @@ impl VerificationEngine {
         Ok(())
     }
 
-    /// Ensure miner-node relationship exists
-    async fn ensure_miner_node_relationship(
-        &self,
-        miner_uid: u16,
-        node_id: &str,
-        node_ssh_endpoint: &str,
-        miner_info: &super::types::MinerInfo,
-    ) -> Result<()> {
-        info!(
-            miner_uid = miner_uid,
-            node_id = node_id,
-            "Ensuring miner-node relationship for miner {} and node {} with real data",
-            miner_uid,
-            node_id
-        );
-
-        let miner_id = format!("miner_{miner_uid}");
-
-        // First ensure the miner exists in miners table with real data
-        self.ensure_miner_exists_with_info(miner_info).await?;
-
-        // Check if relationship already exists
-        let query = "SELECT COUNT(*) as count FROM miner_nodes WHERE miner_id = ? AND node_id = ?";
-        let row = sqlx::query(query)
-            .bind(&miner_id)
-            .bind(node_id)
-            .fetch_one(self.persistence.pool())
-            .await
-            .map_err(|e| anyhow::anyhow!("Failed to check miner-node relationship: {}", e))?;
-
-        let count: i64 = row.get("count");
-
-        if count == 0 {
-            // Check if this node_ssh_endpoint is already used by a different miner
-            let existing_miner: Option<String> = sqlx::query_scalar(
-                "SELECT miner_id FROM miner_nodes WHERE node_ssh_endpoint = ? AND miner_id != ? LIMIT 1",
-            )
-            .bind(node_ssh_endpoint)
-            .bind(&miner_id)
-            .fetch_optional(self.persistence.pool())
-            .await
-            .map_err(|e| anyhow::anyhow!("Failed to check node_ssh_endpoint uniqueness: {}", e))?;
-
-            if let Some(other_miner) = existing_miner {
-                return Err(anyhow::anyhow!(
-                    "Cannot create node relationship: node_ssh_endpoint {} is already registered to {}",
-                    node_ssh_endpoint,
-                    other_miner
-                ));
-            }
-
-            // Check if this is an node ID change for the same miner
-            let old_node_id: Option<String> = sqlx::query_scalar(
-                "SELECT node_id FROM miner_nodes WHERE node_ssh_endpoint = ? AND miner_id = ?",
-            )
-            .bind(node_ssh_endpoint)
-            .bind(&miner_id)
-            .fetch_optional(self.persistence.pool())
-            .await
-            .map_err(|e| anyhow::anyhow!("Failed to check for existing node: {}", e))?;
-
-            if let Some(old_id) = old_node_id {
-                info!(
-                    "Miner {} is changing node ID from {} to {} for endpoint {}",
-                    miner_id, old_id, node_id, node_ssh_endpoint
-                );
-
-                let mut tx = self.persistence.pool().begin().await?;
-
-                sqlx::query(
-                    "UPDATE gpu_uuid_assignments SET node_id = ? WHERE node_id = ? AND miner_id = ?"
-                )
-                .bind(node_id)
-                .bind(&old_id)
-                .bind(&miner_id)
-                .execute(&mut *tx)
-                .await?;
-
-                sqlx::query("DELETE FROM miner_nodes WHERE node_id = ? AND miner_id = ?")
-                    .bind(&old_id)
-                    .bind(&miner_id)
-                    .execute(&mut *tx)
-                    .await?;
-
-                tx.commit().await?;
-
-                info!(
-                    "Successfully migrated GPU assignments from node {} to {}",
-                    old_id, node_id
-                );
-            }
-
-            // Insert new relationship with required fields
-            let insert_query = r#"
-                INSERT OR IGNORE INTO miner_nodes (
-                    id, miner_id, node_id, node_ssh_endpoint, gpu_count,
-                    status, created_at, updated_at
-                ) VALUES (?, ?, ?, ?, ?, ?, datetime('now'), datetime('now'))
-            "#;
-
-            let relationship_id = format!("{miner_id}_{node_id}");
-
-            sqlx::query(insert_query)
-                .bind(&relationship_id)
-                .bind(&miner_id)
-                .bind(node_id)
-                .bind(node_ssh_endpoint)
-                // -- these will be updated from verification details
-                .bind(0) // gpu_count
-                //---------
-                .bind("online") // status - online until verification completes
-                .execute(self.persistence.pool())
-                .await
-                .map_err(|e| anyhow::anyhow!("Failed to insert miner-node relationship: {}", e))?;
-
-            info!(
-                miner_uid = miner_uid,
-                node_id = node_id,
-                "Created miner-node relationship: {} -> {} with endpoint {}",
-                miner_id,
-                node_id,
-                node_ssh_endpoint
-            );
-        } else {
-            debug!(
-                miner_uid = miner_uid,
-                node_id = node_id,
-                "Miner-node relationship already exists: {} -> {}",
-                miner_id,
-                node_id
-            );
-
-            // Even if relationship exists, check for duplicates with same node_ssh_endpoint
-            let duplicate_check_query: &'static str =
-                "SELECT id, node_id FROM miner_nodes WHERE node_ssh_endpoint = ? AND id != ?";
-            let relationship_id = format!("{miner_id}_{node_id}");
-
-            let duplicates = sqlx::query(duplicate_check_query)
-                .bind(node_ssh_endpoint)
-                .bind(&relationship_id)
-                .fetch_all(self.persistence.pool())
-                .await
-                .map_err(|e| anyhow::anyhow!("Failed to check for duplicate nodes: {}", e))?;
-
-            if !duplicates.is_empty() {
-                let duplicate_count = duplicates.len();
-                warn!(
-                    miner_uid = miner_uid,
-                    "Found {} duplicate nodes with same node_ssh_endpoint {} for miner {}",
-                    duplicate_count,
-                    node_ssh_endpoint,
-                    miner_id
-                );
-
-                // Delete the duplicates to clean up fraudulent registrations
-                for duplicate in duplicates {
-                    let dup_id: String = duplicate.get("id");
-                    let dup_node_id: String = duplicate.get("node_id");
-
-                    warn!(
-                        miner_uid = miner_uid,
-                        "Marking duplicate node {} (id: {}) as offline with same node_ssh_endpoint as {} for miner {}",
-                        dup_node_id, dup_id, node_id, miner_id
-                    );
-
-                    sqlx::query("UPDATE miner_nodes SET status = 'offline', last_health_check = datetime('now'), updated_at = datetime('now') WHERE id = ?")
-                        .bind(&dup_id)
-                        .execute(self.persistence.pool())
-                        .await
-                        .map_err(|e| {
-                            anyhow::anyhow!("Failed to update duplicate node status: {}", e)
-                        })?;
-
-                    // Also clean up associated GPU assignments for the duplicate
-                    self.cleanup_gpu_assignments(&dup_node_id, &miner_id, None)
-                        .await
-                        .map_err(|e| {
-                            anyhow::anyhow!(
-                                "Failed to clean up GPU assignments for duplicate: {}",
-                                e
-                            )
-                        })?;
-                }
-
-                info!(
-                    miner_uid = miner_uid,
-                    "Cleaned up {} duplicate nodes for miner {} with node_ssh_endpoint {}",
-                    duplicate_count,
-                    miner_id,
-                    node_ssh_endpoint
-                );
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Store GPU UUID assignments for an node
-    async fn store_gpu_uuid_assignments(
-        &self,
-        miner_uid: u16,
-        node_id: &str,
-        gpu_infos: &[GpuInfo],
-    ) -> Result<()> {
-        let miner_id = format!("miner_{miner_uid}");
-        let now = chrono::Utc::now().to_rfc3339();
-
-        // Collect all valid GPU UUIDs being reported
-        let reported_gpu_uuids: Vec<String> = gpu_infos
-            .iter()
-            .filter(|g| !g.gpu_uuid.is_empty() && g.gpu_uuid != "Unknown UUID")
-            .map(|g| g.gpu_uuid.clone())
-            .collect();
-
-        // Clean up GPU assignments based on what's reported
-        if !reported_gpu_uuids.is_empty() {
-            // Some GPUs reported - clean up any that are no longer reported
-            let placeholders = reported_gpu_uuids
-                .iter()
-                .map(|_| "?")
-                .collect::<Vec<_>>()
-                .join(", ");
-            let query = format!(
-                "DELETE FROM gpu_uuid_assignments
-                 WHERE miner_id = ? AND node_id = ?
-                 AND gpu_uuid NOT IN ({placeholders})"
-            );
-
-            let mut q = sqlx::query(&query).bind(&miner_id).bind(node_id);
-
-            for uuid in &reported_gpu_uuids {
-                q = q.bind(uuid);
-            }
-
-            let deleted = q.execute(self.persistence.pool()).await?;
-
-            if deleted.rows_affected() > 0 {
-                info!(
-                    miner_uid = miner_uid,
-                    "Cleaned up {} stale GPU assignments for {}/{}",
-                    deleted.rows_affected(),
-                    miner_id,
-                    node_id
-                );
-            }
-        } else {
-            // No GPUs reported - clean up all assignments for this node
-            let deleted_rows = self
-                .cleanup_gpu_assignments(node_id, &miner_id, None)
-                .await?;
-
-            if deleted_rows > 0 {
-                info!(
-                    miner_uid = miner_uid,
-                    "Cleaned up {} GPU assignments for {}/{} (no GPUs reported)",
-                    deleted_rows,
-                    miner_id,
-                    node_id
-                );
-            }
-        }
-
-        for gpu_info in gpu_infos {
-            // Skip invalid UUIDs
-            if gpu_info.gpu_uuid.is_empty() || gpu_info.gpu_uuid == "Unknown UUID" {
-                continue;
-            }
-
-            // Check if this GPU UUID already exists
-            let existing = sqlx::query(
-                "SELECT miner_id, node_id FROM gpu_uuid_assignments WHERE gpu_uuid = ?",
-            )
-            .bind(&gpu_info.gpu_uuid)
-            .fetch_optional(self.persistence.pool())
-            .await?;
-
-            if let Some(row) = existing {
-                let existing_miner_id: String = row.get("miner_id");
-                let existing_node_id: String = row.get("node_id");
-
-                if existing_miner_id != miner_id || existing_node_id != node_id {
-                    // Check if the existing node is still active
-                    let node_status_query =
-                        "SELECT status FROM miner_nodes WHERE node_id = ? AND miner_id = ?";
-                    let status_row = sqlx::query(node_status_query)
-                        .bind(&existing_node_id)
-                        .bind(&existing_miner_id)
-                        .fetch_optional(self.persistence.pool())
-                        .await?;
-
-                    let can_reassign = if let Some(row) = status_row {
-                        let status: String = row.get("status");
-                        // Allow reassignment if node is offline, failed, or stale
-                        status == "offline" || status == "failed" || status == "stale"
-                    } else {
-                        // Node doesn't exist in miner_nodes table - allow reassignment
-                        true
-                    };
-
-                    if can_reassign {
-                        // GPU reassignment allowed - previous node is inactive
-                        info!(
-                            miner_uid = miner_uid,
-                            security = true,
-                            gpu_uuid = %gpu_info.gpu_uuid,
-                            previous_miner_id = %existing_miner_id,
-                            previous_node_id = %existing_node_id,
-                            new_miner_id = %miner_id,
-                            new_node_id = %node_id,
-                            gpu_memory_gb = %gpu_info.gpu_memory_gb,
-                            action = "gpu_assignment_reassigned",
-                            reassignment_reason = "previous_node_inactive",
-                            "GPU {} reassigned from {}/{} to {}/{} (previous node inactive)",
-                            gpu_info.gpu_uuid,
-                            existing_miner_id,
-                            existing_node_id,
-                            miner_id,
-                            node_id
-                        );
-
-                        sqlx::query(
-                            "UPDATE gpu_uuid_assignments
-                             SET miner_id = ?, node_id = ?, gpu_index = ?, gpu_name = ?,
-                                 gpu_memory_gb = ?, last_verified = ?, updated_at = ?
-                             WHERE gpu_uuid = ?",
-                        )
-                        .bind(&miner_id)
-                        .bind(node_id)
-                        .bind(gpu_info.index as i32)
-                        .bind(&gpu_info.gpu_name)
-                        .bind(gpu_info.gpu_memory_gb)
-                        .bind(&now)
-                        .bind(&now)
-                        .bind(&gpu_info.gpu_uuid)
-                        .execute(self.persistence.pool())
-                        .await?;
-                    } else {
-                        // Node is still active - reject the reassignment
-                        warn!(
-                            miner_uid = miner_uid,
-                            security = true,
-                            gpu_uuid = %gpu_info.gpu_uuid,
-                            existing_miner_id = %existing_miner_id,
-                            existing_node_id = %existing_node_id,
-                            attempting_miner_id = %miner_id,
-                            attempting_node_id = %node_id,
-                            action = "gpu_assignment_rejected",
-                            rejection_reason = "already_owned_by_active_node",
-                            "GPU UUID {} still owned by active node {}/{}, rejecting claim from {}/{}",
-                            gpu_info.gpu_uuid,
-                            existing_miner_id,
-                            existing_node_id,
-                            miner_id,
-                            node_id
-                        );
-                        // Skip this GPU - don't store it for the new claimant
-                        continue;
-                    }
-                } else {
-                    // Same owner - just update last_verified
-                    sqlx::query(
-                        "UPDATE gpu_uuid_assignments
-                         SET gpu_memory_gb = ?, last_verified = ?, updated_at = ?
-                         WHERE gpu_uuid = ?",
-                    )
-                    .bind(gpu_info.gpu_memory_gb)
-                    .bind(&now)
-                    .bind(&now)
-                    .bind(&gpu_info.gpu_uuid)
-                    .execute(self.persistence.pool())
-                    .await?;
-                }
-            } else {
-                // New GPU UUID - insert
-                sqlx::query(
-                    "INSERT INTO gpu_uuid_assignments
-                     (gpu_uuid, gpu_index, node_id, miner_id, gpu_name, gpu_memory_gb, last_verified, created_at, updated_at)
-                     VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)"
-                )
-                .bind(&gpu_info.gpu_uuid)
-                .bind(gpu_info.index as i32)
-                .bind(node_id)
-                .bind(&miner_id)
-                .bind(&gpu_info.gpu_name)
-                .bind(gpu_info.gpu_memory_gb)
-                .bind(&now)
-                .bind(&now)
-                .bind(&now)
-                .execute(self.persistence.pool())
-                .await?;
-
-                info!(
-                    miner_uid = miner_uid,
-                    security = true,
-                    gpu_uuid = %gpu_info.gpu_uuid,
-                    gpu_index = gpu_info.index,
-                    node_id = %node_id,
-                    miner_id = %miner_id,
-                    gpu_name = %gpu_info.gpu_name,
-                    gpu_memory_gb = %gpu_info.gpu_memory_gb,
-                    action = "gpu_assignment_created",
-                    "Registered new GPU {} (index {}) for {}/{}",
-                    gpu_info.gpu_uuid, gpu_info.index, miner_id, node_id
-                );
-            }
-        }
-
-        // Update gpu_count in miner_nodes based on actual GPU assignments
-        let gpu_count = sqlx::query_scalar::<_, i64>(
-            "SELECT COUNT(*) FROM gpu_uuid_assignments WHERE miner_id = ? AND node_id = ?",
-        )
-        .bind(&miner_id)
-        .bind(node_id)
-        .fetch_one(self.persistence.pool())
-        .await?;
-
-        // Status hierarchy: "online" > "verified" > "offline"
-        let current_status = sqlx::query_scalar::<_, String>(
-            "SELECT status FROM miner_nodes WHERE miner_id = ? AND node_id = ?",
-        )
-        .bind(&miner_id)
-        .bind(node_id)
-        .fetch_one(self.persistence.pool())
-        .await?;
-
-        let new_status = match (current_status.as_str(), gpu_count > 0) {
-            ("online", true) => "online",   // Keep online status if GPUs present
-            ("verified", true) => "online", // Promote verified back to online if GPUs present
-            ("online", false) => "offline", // Downgrade to offline if no GPUs
-            (_, true) => "verified",        // Set verified if GPUs present and not online
-            (_, false) => "offline",        // Set offline if no GPUs
-        };
-
-        sqlx::query(
-            "UPDATE miner_nodes SET gpu_count = ?, status = ?, updated_at = datetime('now')
-             WHERE miner_id = ? AND node_id = ?",
-        )
-        .bind(gpu_count as i32)
-        .bind(new_status)
-        .bind(&miner_id)
-        .bind(node_id)
-        .execute(self.persistence.pool())
-        .await?;
-
-        if gpu_count > 0 {
-            info!(
-                security = true,
-                node_id = %node_id,
-                miner_id = %miner_id,
-                gpu_count = gpu_count,
-                new_status = %new_status,
-                action = "node_gpu_verification_success",
-                "Node {}/{} verified with {} GPUs, status: {}",
-                miner_id, node_id, gpu_count, new_status
-            );
-        } else {
-            warn!(
-                miner_uid = miner_uid,
-                security = true,
-                node_id = %node_id,
-                miner_id = %miner_id,
-                gpu_count = 0,
-                new_status = %new_status,
-                action = "node_gpu_verification_failure",
-                "Node {}/{} has no GPUs, marking as {}",
-                miner_id, node_id, new_status
-            );
-        }
-
-        // Validate that the GPU count matches the expected count
-        let expected_gpu_count = gpu_infos
-            .iter()
-            .filter(|g| !g.gpu_uuid.is_empty() && g.gpu_uuid != "Unknown UUID")
-            .count() as i64;
-
-        if gpu_count != expected_gpu_count {
-            warn!(
-                miner_uid = miner_uid,
-                "GPU assignment mismatch for {}/{}: stored {} GPUs but expected {}",
-                miner_id,
-                node_id,
-                gpu_count,
-                expected_gpu_count
-            );
-        }
-
-        // Fail verification if node claims GPUs but none were stored
-        if expected_gpu_count > 0 && gpu_count == 0 {
-            error!(
-                miner_uid = miner_uid,
-                "Failed to store GPU assignments for {}/{}: expected {} GPUs but stored 0",
-                miner_id,
-                node_id,
-                expected_gpu_count
-            );
-            return Err(anyhow::anyhow!(
-                "GPU assignment validation failed: no valid GPU UUIDs stored despite {} GPUs reported",
-                expected_gpu_count
-            ));
-        }
-
-        Ok(())
-    }
-
-    /// Update last_verified timestamp for existing GPU assignments
-    async fn update_gpu_assignment_timestamps(
-        &self,
-        miner_uid: u16,
-        node_id: &str,
-        gpu_infos: &[GpuInfo],
-    ) -> Result<()> {
-        let miner_id = format!("miner_{miner_uid}");
-        let now = chrono::Utc::now().to_rfc3339();
-
-        let reported_gpu_uuids: Vec<String> = gpu_infos
-            .iter()
-            .filter(|g| !g.gpu_uuid.is_empty() && g.gpu_uuid != "Unknown UUID")
-            .map(|g| g.gpu_uuid.clone())
-            .collect();
-
-        if reported_gpu_uuids.is_empty() {
-            debug!(
-                miner_uid = miner_uid,
-                "No valid GPU UUIDs reported for {}/{} in lightweight validation",
-                miner_id,
-                node_id
-            );
-            return Ok(());
-        }
-
-        let placeholders = reported_gpu_uuids
-            .iter()
-            .map(|_| "?")
-            .collect::<Vec<_>>()
-            .join(", ");
-
-        let query = format!(
-            "UPDATE gpu_uuid_assignments
-             SET last_verified = ?, updated_at = ?
-             WHERE miner_id = ? AND node_id = ? AND gpu_uuid IN ({placeholders})"
-        );
-
-        let mut q = sqlx::query(&query)
-            .bind(&now)
-            .bind(&now)
-            .bind(&miner_id)
-            .bind(node_id);
-
-        for uuid in &reported_gpu_uuids {
-            q = q.bind(uuid);
-        }
-
-        let result = q.execute(self.persistence.pool()).await?;
-        let updated_count = result.rows_affected();
-
-        if updated_count > 0 {
-            info!(
-                security = true,
-                miner_uid = miner_uid,
-                node_id = %node_id,
-                validation_type = "lightweight",
-                updated_assignments = updated_count,
-                action = "gpu_assignment_timestamp_updated",
-                "Updated {} GPU assignment timestamps for {}/{} (lightweight validation)",
-                updated_count, miner_id, node_id
-            );
-        } else {
-            debug!(
-                security = true,
-                miner_uid = miner_uid,
-                node_id = %node_id,
-                validation_type = "lightweight",
-                "No GPU assignments found to update for {}/{} with {} reported UUIDs",
-                miner_id,
-                node_id,
-                reported_gpu_uuids.len()
-            );
-            if self
-                .persistence
-                .has_active_rental(node_id, &miner_id)
-                .await
-                .unwrap_or(false)
-            {
-                self.store_gpu_uuid_assignments(miner_uid, node_id, gpu_infos)
-                    .await?;
-            } else {
-                debug!(
-                    security = true,
-                    miner_uid = miner_uid,
-                    node_id = %node_id,
-                    validation_type = "lightweight",
-                    "Skipping GPU assignment creation in lightweight (no active rental)"
-                );
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Ensure miner exists in miners table
-    ///
-    /// This function handles three scenarios:
-    /// 1. if UID already exists with same hotkey -> Update data
-    /// 2. if UID already exists with different hotkey -> Update to new hotkey (recycled UID)
-    /// 3. if UID doesn't exist but hotkey does -> on re-registration, migrate the UID
-    /// 4. if neither UID nor hotkey exist -> Create new miner
-    async fn ensure_miner_exists_with_info(
-        &self,
-        miner_info: &super::types::MinerInfo,
-    ) -> Result<()> {
-        let new_miner_uid = format!("miner_{}", miner_info.uid.as_u16());
-        let hotkey = miner_info.hotkey.to_string();
-
-        // Step 1: handle recycled UIDs
-        let existing_by_uid = self.check_miner_by_uid(&new_miner_uid).await?;
-
-        if let Some((_, existing_hotkey)) = existing_by_uid {
-            return self
-                .handle_recycled_miner_uid(&new_miner_uid, &hotkey, &existing_hotkey, miner_info)
-                .await;
-        }
-
-        // Step 2: handle UID changes when a hotkey moves to a new UID (re-registration)
-        let existing_by_hotkey = self.check_miner_by_hotkey(&hotkey).await?;
-
-        if let Some(old_miner_uid) = existing_by_hotkey {
-            return self
-                .handle_uid_change(&old_miner_uid, &new_miner_uid, &hotkey, miner_info)
-                .await;
-        }
-
-        // Step 3: handle new miners when neither UID nor hotkey exist - create new miner
-        self.create_new_miner(&new_miner_uid, &hotkey, miner_info)
-            .await
-    }
-
-    /// Check if a miner with the given UID exists
-    async fn check_miner_by_uid(&self, miner_uid: &str) -> Result<Option<(String, String)>> {
-        let query = "SELECT id, hotkey FROM miners WHERE id = ?";
-        let result = sqlx::query(query)
-            .bind(miner_uid)
-            .fetch_optional(self.persistence.pool())
-            .await
-            .map_err(|e| anyhow::anyhow!("Failed to check miner by uid: {}", e))?;
-
-        Ok(result.map(|row| {
-            let id: String = row.get("id");
-            let hotkey: String = row.get("hotkey");
-            (id, hotkey)
-        }))
-    }
-
-    /// Check if a miner with the given hotkey exists
-    async fn check_miner_by_hotkey(&self, hotkey: &str) -> Result<Option<String>> {
-        let query = "SELECT id FROM miners WHERE hotkey = ?";
-        let result = sqlx::query(query)
-            .bind(hotkey)
-            .fetch_optional(self.persistence.pool())
-            .await
-            .map_err(|e| anyhow::anyhow!("Failed to check miner by hotkey: {}", e))?;
-
-        Ok(result.map(|row| row.get("id")))
-    }
-
-    /// Handle case where miner UID already exists
-    async fn handle_recycled_miner_uid(
-        &self,
-        miner_uid: &str,
-        new_hotkey: &str,
-        existing_hotkey: &str,
-        miner_info: &super::types::MinerInfo,
-    ) -> Result<()> {
-        if existing_hotkey != new_hotkey {
-            // Case: Recycled UID - same UID but different hotkey
-            info!(
-                miner_uid = miner_uid,
-                "Miner {} exists with old hotkey {}, updating to new hotkey {}",
-                miner_uid,
-                existing_hotkey,
-                new_hotkey
-            );
-
-            let update_query = r#"
-                UPDATE miners SET
-                    hotkey = ?, endpoint = ?, verification_score = ?,
-                    last_seen = datetime('now'), updated_at = datetime('now')
-                WHERE id = ?
-            "#;
-
-            sqlx::query(update_query)
-                .bind(new_hotkey)
-                .bind(&miner_info.endpoint)
-                .bind(miner_info.verification_score)
-                .bind(miner_uid)
-                .execute(self.persistence.pool())
-                .await
-                .map_err(|e| anyhow::anyhow!("Failed to update miner with new hotkey: {}", e))?;
-
-            debug!("Updated miner {} with new hotkey and data", miner_uid);
-        } else {
-            // Case: Same miner, same hotkey - just update the data
-            self.update_miner_data(miner_uid, miner_info).await?;
-        }
-
-        Ok(())
-    }
-
-    /// Handle case where hotkey exists but with different ID (UID change)
-    async fn handle_uid_change(
-        &self,
-        old_miner_id: &str,
-        new_miner_id: &str,
-        hotkey: &str,
-        miner_info: &super::types::MinerInfo,
-    ) -> Result<()> {
-        info!(
-            "Detected UID change for hotkey {}: {} -> {}",
-            hotkey, old_miner_id, new_miner_id
-        );
-
-        // Migrate the miner UID
-        if let Err(e) = self
-            .migrate_miner_uid(old_miner_id, new_miner_id, miner_info)
-            .await
-        {
-            error!(
-                "Failed to migrate miner UID from {} to {}: {}",
-                old_miner_id, new_miner_id, e
-            );
-            return Err(e);
-        }
-
-        Ok(())
-    }
-
-    /// Update existing miner data
-    async fn update_miner_data(
-        &self,
-        miner_id: &str,
-        miner_info: &super::types::MinerInfo,
-    ) -> Result<()> {
-        let update_query = r#"
-            UPDATE miners SET
-                endpoint = ?, verification_score = ?,
-                last_seen = datetime('now'), updated_at = datetime('now')
-            WHERE id = ?
-        "#;
-
-        sqlx::query(update_query)
-            .bind(&miner_info.endpoint)
-            .bind(miner_info.verification_score)
-            .bind(miner_id)
-            .execute(self.persistence.pool())
-            .await
-            .map_err(|e| anyhow::anyhow!("Failed to update miner: {}", e))?;
-
-        debug!("Updated miner record: {} with latest data", miner_id);
-        Ok(())
-    }
-
-    /// Create a new miner record
-    async fn create_new_miner(
-        &self,
-        miner_uid: &str,
-        hotkey: &str,
-        miner_info: &super::types::MinerInfo,
-    ) -> Result<()> {
-        let insert_query = r#"
-            INSERT INTO miners (
-                id, hotkey, endpoint, verification_score, uptime_percentage,
-                last_seen, registered_at, updated_at, node_info
-            ) VALUES (?, ?, ?, ?, ?, datetime('now'), datetime('now'), datetime('now'), ?)
-        "#;
-
-        sqlx::query(insert_query)
-            .bind(miner_uid)
-            .bind(hotkey)
-            .bind(&miner_info.endpoint)
-            .bind(miner_info.verification_score)
-            .bind(100.0) // uptime_percentage
-            .bind("{}") // node_info
-            .execute(self.persistence.pool())
-            .await
-            .map_err(|e| anyhow::anyhow!("Failed to insert miner: {}", e))?;
-
-        info!(
-            "Created miner record: {} with hotkey {} and endpoint {}",
-            miner_uid, hotkey, miner_info.endpoint
-        );
-
-        Ok(())
-    }
-
-    /// Migrate miner UID when it changes in the network
-    async fn migrate_miner_uid(
-        &self,
-        old_miner_uid: &str,
-        new_miner_uid: &str,
-        miner_info: &super::types::MinerInfo,
-    ) -> Result<()> {
-        info!(
-            "Starting UID migration: {} -> {} for hotkey {}",
-            old_miner_uid, new_miner_uid, miner_info.hotkey
-        );
-
-        // Use a transaction to ensure atomicity
-        let mut tx = self
-            .persistence
-            .pool()
-            .begin()
-            .await
-            .map_err(|e| anyhow::anyhow!("Failed to begin transaction: {}", e))?;
-
-        // 1. First, get the old miner data
-        debug!("Fetching old miner record: {}", old_miner_uid);
-        let get_old_miner = "SELECT * FROM miners WHERE id = ?";
-        let old_miner_row = sqlx::query(get_old_miner)
-            .bind(old_miner_uid)
-            .fetch_optional(&mut *tx)
-            .await
-            .map_err(|e| anyhow::anyhow!("Failed to fetch old miner record: {}", e))?;
-
-        if old_miner_row.is_none() {
-            return Err(anyhow::anyhow!(
-                "Old miner record not found: {}",
-                old_miner_uid
-            ));
-        }
-
-        let old_row = old_miner_row.unwrap();
-        debug!("Found old miner record for migration");
-
-        // 2. Check if any miner with this hotkey exists (including the target)
-        debug!(
-            "Checking for existing miners with hotkey: {}",
-            miner_info.hotkey
-        );
-        let check_hotkey = "SELECT id FROM miners WHERE hotkey = ?";
-        let all_with_hotkey = sqlx::query(check_hotkey)
-            .bind(miner_info.hotkey.to_string())
-            .fetch_all(&mut *tx)
-            .await
-            .map_err(|e| anyhow::anyhow!("Failed to check hotkey existence: {}", e))?;
-
-        // Find if any of them is NOT the old miner
-        let existing_with_hotkey = all_with_hotkey.into_iter().find(|row| {
-            let id: String = row.get("id");
-            id != old_miner_uid
-        });
-
-        let should_create_new = if let Some(row) = existing_with_hotkey {
-            let existing_id: String = row.get("id");
-            debug!(
-                "Found existing miner with hotkey {}: id={}",
-                miner_info.hotkey, existing_id
-            );
-            if existing_id == new_miner_uid {
-                // The new miner record already exists, just need to delete old
-                debug!("New miner record already exists with correct ID");
-                false
-            } else {
-                // Another miner exists with this hotkey but different ID
-                warn!(
-                    "Cannot migrate: Another miner {} already exists with hotkey {} (trying to create {})",
-                    existing_id, miner_info.hotkey, new_miner_uid
-                );
-                return Err(anyhow::anyhow!(
-                    "Cannot migrate: Another miner {} already exists with hotkey {}",
-                    existing_id,
-                    miner_info.hotkey
-                ));
-            }
-        } else {
-            debug!(
-                "No existing miner with hotkey {}, will create new record",
-                miner_info.hotkey
-            );
-            true
-        };
-
-        // Extract old miner data we'll need
-        let verification_score = old_row
-            .try_get::<f64, _>("verification_score")
-            .unwrap_or(0.0);
-        let uptime_percentage = old_row
-            .try_get::<f64, _>("uptime_percentage")
-            .unwrap_or(100.0);
-        let registered_at = old_row
-            .try_get::<String, _>("registered_at")
-            .unwrap_or_else(|_| chrono::Utc::now().to_rfc3339());
-        let node_info = old_row
-            .try_get::<String, _>("node_info")
-            .unwrap_or_else(|_| "{}".to_string());
-
-        // 3. Get all related data before deletion
-        debug!("Fetching related node data");
-        let get_nodes = "SELECT * FROM miner_nodes WHERE miner_id = ?";
-        let nodes = sqlx::query(get_nodes)
-            .bind(old_miner_uid)
-            .fetch_all(&mut *tx)
-            .await
-            .map_err(|e| anyhow::anyhow!("Failed to fetch nodes: {}", e))?;
-
-        debug!("Found {} nodes to migrate", nodes.len());
-
-        // 4. Delete old miner record (this will CASCADE delete miner_nodes and verification_requests)
-        debug!("Deleting old miner record: {}", old_miner_uid);
-        let delete_old_miner = "DELETE FROM miners WHERE id = ?";
-        sqlx::query(delete_old_miner)
-            .bind(old_miner_uid)
-            .execute(&mut *tx)
-            .await
-            .map_err(|e| anyhow::anyhow!("Failed to delete old miner record: {}", e))?;
-
-        debug!("Deleted old miner record and related data");
-
-        // 5. Create new miner record if needed
-        if should_create_new {
-            debug!("Creating new miner record: {}", new_miner_uid);
-            let insert_new_miner = r#"
-                INSERT INTO miners (
-                    id, hotkey, endpoint, verification_score, uptime_percentage,
-                    last_seen, registered_at, updated_at, node_info
-                ) VALUES (?, ?, ?, ?, ?, datetime('now'), ?, datetime('now'), ?)
-            "#;
-
-            sqlx::query(insert_new_miner)
-                .bind(new_miner_uid)
-                .bind(miner_info.hotkey.to_string())
-                .bind(&miner_info.endpoint)
-                .bind(verification_score)
-                .bind(uptime_percentage)
-                .bind(registered_at)
-                .bind(node_info)
-                .execute(&mut *tx)
-                .await
-                .map_err(|e| anyhow::anyhow!("Failed to create new miner record: {}", e))?;
-
-            debug!("Successfully created new miner record");
-        }
-
-        // 6. Re-create node relationships
-        let mut node_count = 0;
-        for node_row in nodes {
-            let node_id: String = node_row.get("node_id");
-            let ssh_endpoint: String = node_row.get("ssh_endpoint");
-            let gpu_count: i32 = node_row.get("gpu_count");
-            let status: String = node_row
-                .try_get("status")
-                .unwrap_or_else(|_| "unknown".to_string());
-            // Check if this ssh_endpoint is already in use by another miner
-            let existing_check = sqlx::query(
-                "SELECT COUNT(*) as count FROM miner_nodes WHERE ssh_endpoint = ? AND miner_id != ?"
-            )
-            .bind(&ssh_endpoint)
-            .bind(new_miner_uid)
-            .fetch_one(&mut *tx)
-            .await?;
-
-            let existing_count: i64 = existing_check.get("count");
-            if existing_count > 0 {
-                warn!(
-                    "Skipping node {} during UID migration: ssh_endpoint {} already in use by another miner",
-                    node_id, ssh_endpoint
-                );
-                continue;
-            }
-
-            let new_id = format!("{new_miner_uid}_{node_id}");
-
-            let insert_node = r#"
-                INSERT INTO miner_nodes (
-                    id, miner_id, node_id, ssh_endpoint, gpu_count,
-                    status, last_health_check,
-                    created_at, updated_at
-                ) VALUES (?, ?, ?, ?, ?, ?, NULL, datetime('now'), datetime('now'))
-            "#;
-
-            sqlx::query(insert_node)
-                .bind(&new_id)
-                .bind(new_miner_uid)
-                .bind(&node_id)
-                .bind(&ssh_endpoint)
-                .bind(gpu_count)
-                .bind(&status)
-                .execute(&mut *tx)
-                .await
-                .map_err(|e| anyhow::anyhow!("Failed to recreate node relationship: {}", e))?;
-
-            node_count += 1;
-        }
-
-        debug!("Recreated {} node relationships", node_count);
-
-        // 7. Migrate GPU UUID assignments
-        debug!(
-            "Migrating GPU UUID assignments from {} to {}",
-            old_miner_uid, new_miner_uid
-        );
-        let update_gpu_assignments = r#"
-            UPDATE gpu_uuid_assignments
-            SET miner_id = ?
-            WHERE miner_id = ?
-        "#;
-
-        let gpu_result = sqlx::query(update_gpu_assignments)
-            .bind(new_miner_uid)
-            .bind(old_miner_uid)
-            .execute(&mut *tx)
-            .await?;
-
-        debug!(
-            "Migrated {} GPU UUID assignments",
-            gpu_result.rows_affected()
-        );
-
-        // Commit the transaction
-        debug!("Committing transaction");
-        tx.commit()
-            .await
-            .map_err(|e| anyhow::anyhow!("Failed to commit transaction: {}", e))?;
-
-        info!(
-            "Successfully migrated miner UID: {} -> {}. Migrated {} nodes",
-            old_miner_uid, new_miner_uid, node_count
-        );
-
-        Ok(())
-    }
-
     /// Sync miners from metagraph to database
     pub async fn sync_miners_from_metagraph(&self, miners: &[MinerInfo]) -> Result<()> {
-        info!("Syncing {} miners from metagraph to database", miners.len());
-
-        for miner in miners {
-            // Discovery already filters out miners without valid axon endpoints
-            if let Err(e) = self.ensure_miner_exists_with_info(miner).await {
-                warn!(
-                    "Failed to sync miner {} to database: {}",
-                    miner.uid.as_u16(),
-                    e
-                );
-            } else {
-                debug!(
-                    "Successfully synced miner {} with endpoint {} to database",
-                    miner.uid.as_u16(),
-                    miner.endpoint
-                );
-            }
-        }
-
-        info!("Completed syncing miners from metagraph");
-        Ok(())
+        self.persistence.sync_miners_from_metagraph(miners).await
     }
 
     /// Create authenticated miner client
@@ -2151,6 +1070,22 @@ impl VerificationEngine {
         &self,
         consecutive_failures_threshold: i32,
     ) -> Result<()> {
+        self.persistence
+            .cleanup_failed_nodes_after_failures(
+                consecutive_failures_threshold,
+                self.config.gpu_assignment_cleanup_ttl,
+            )
+            .await?;
+
+        Ok(())
+    }
+
+    /// Original cleanup implementation (now moved to persistence layer)
+    #[allow(dead_code)]
+    async fn cleanup_failed_nodes_after_failures_original(
+        &self,
+        consecutive_failures_threshold: i32,
+    ) -> Result<()> {
         info!(
             "Running node cleanup - checking for {} consecutive failures",
             consecutive_failures_threshold
@@ -2181,6 +1116,7 @@ impl VerificationEngine {
             );
 
             let rows_cleaned = self
+                .persistence
                 .cleanup_gpu_assignments(&node_id, &miner_id, None)
                 .await?;
             gpu_assignments_cleaned += rows_cleaned;
@@ -2321,6 +1257,7 @@ impl VerificationEngine {
             );
 
             let cleaned = self
+                .persistence
                 .cleanup_gpu_assignments(&node_id, &miner_id, None)
                 .await?;
             stale_gpu_cleaned += cleaned;
@@ -2386,7 +1323,8 @@ impl VerificationEngine {
             let mut tx = self.persistence.pool().begin().await?;
 
             // Clean up any remaining GPU assignments
-            self.cleanup_gpu_assignments(&node_id, &miner_id, Some(&mut tx))
+            self.persistence
+                .cleanup_gpu_assignments(&node_id, &miner_id, Some(&mut tx))
                 .await?;
 
             // Delete the node record
